bash bug discovered
./oo.sh

bash(50440,0x7fff7cfc9300) malloc: *** error for object 0x27fd05bd25ca0: pointer being freed was not allocated
*** set a breakpoint in malloc_error_break to debug
[1]    50440 abort      ./oo.sh

+++ lib/system/03_exception.sh
@@ -2,7 +2,7 @@ alias throw="__EXCEPTION_TYPE__='MANUALLY INVOKED' command_not_found_handle"
 
 command_not_found_handle() {
     # ignore the error from the catch subshell itself
-    if [[ "$*" = '( set -e; trap "Exception.Capture ${LINENO}; " ERR;'* ]]
+    if [[ "$*" = '( set -e;'* ]]
     then
         return 0
     fi
@@ -21,10 +21,26 @@ command_not_found_handle() {
     if [[ $__oo__insideTryCatch -gt 0 ]]
     then
         Log.Debug 3 "inside Try Number $__oo__insideTryCatch"
-        echo "$undefinedObject" > /tmp/stored_exception
-        echo "$lineNo" > /tmp/stored_exception_line
-#        echo "$script" > /tmp/stored_exception_source
-        return 1
+
+        if [[ ! -f /tmp/stored_exception_line ]]; then
+            echo "$lineNo" > /tmp/stored_exception_line
+        fi
+        if [[ ! -f /tmp/stored_exception ]]; then
+            echo "$undefinedObject" > /tmp/stored_exception
+        fi
+        if [[ ! -f /tmp/stored_exception_source ]]; then
+            echo "$script" > /tmp/stored_exception_source
+        fi
+        if [[ ! -f /tmp/stored_exception_backtrace ]]; then
+            Exception.DumpBacktrace 2 > /tmp/stored_exception_backtrace
+        fi
+        #Exception.DumpBacktrace 1
+        
+        #echo "$lineNo"
+        #echo "$undefinedObject"
+        #echo "$lineNo"
+        
+        return 1 # needs to be return 1
     fi
 
     if [[ $BASH_SUBSHELL -ge 25 ]]
@@ -35,8 +51,12 @@ command_not_found_handle() {
 
     Log.Write
     Log.Write " $(UI.Color.Red)$(UI.Powerline.Fail) $(UI.Color.Bold)UNCAUGHT EXCEPTION${merger}$(UI.Color.LightRed)${type}$(UI.Color.Default)"
-    Log.Write "$(Exception.FormatException "$script" "$lineNo" "$undefinedObject")"
-    Log.Write "$(Exception.FormatBacktrace 3 "$script" "$lineNo" "$undefinedObject")"
+    Log.Write "$(Exception.FormatExceptionSegment "$script" "$lineNo" "$undefinedObject")"
+    
+    local -i backtraceLevel=1
+    Exception.PrintException $(IFS=$'\n'; Exception.DumpBacktrace 2)
+    
+    #Log.Write "$(Exception.FormatBacktrace 3 "$script" "$lineNo" "$undefinedObject")"
 
     Exception.ContinueOrBreak
 }
@@ -68,49 +88,148 @@ Exception.FormatBacktrace() {
 
     if [[ "$exceptionFile" != "${traceFile[0]}" || "$exceptionLine" != "${traceLine[0]}" ]]
     then
-        echo "$(Exception.FormatException "${traceFile[0]}" "${traceLine[0]}" "$exceptionCommand" 0)"
+        echo "$(Exception.FormatExceptionSegment "${traceFile[0]}" "${traceLine[0]}" "$exceptionCommand" 0)"
     fi
 
     while [[ $index -lt $traceCount ]]
     do
         # Log.Write "${traceFile[$index]}" \| "${traceLine[$index]}" \| "${traceCommand[($index - 1)]}" \| $(($index + 1))
 
-        echo "$(Exception.FormatException "${traceFile[$index]}" "${traceLine[$index]}" "${traceCommand[($index - 1)]}" $(($index + 1)) )"
+        echo "$(Exception.FormatExceptionSegment "${traceFile[$index]}" "${traceLine[$index]}" "${traceCommand[($index - 1)]}" $(($index + 1)) )"
         index+=1
     done
-    # echo "$(Exception.FormatException "${traceFile[($index - 1)]}" "${traceLine[($index - 1)]}" "${traceCommand[($index - 1)]}" $(($index)) )"
+    # echo "$(Exception.FormatExceptionSegment "${traceFile[($index - 1)]}" "${traceLine[($index - 1)]}" "${traceCommand[($index - 1)]}" $(($index)) )"
 }
 
-Exception.FormatException() {
+Exception.PrintException() {
+    @params exception
+    
+    local -i backtraceLevel=${backtraceLevel:-0}
+    
+    local -i counter=0
+    local -i backtraceNo=0
+    
+    local -a backtraceLine
+    local -a backtraceCommand
+    local -a backtraceFile
+    
+    #for traceElement in Exception.GetLastException
+    while [[ $counter -lt ${#exception[@]} ]]
+    do
+        backtraceLine[$backtraceNo]="${exception[$counter]}"
+        counter+=1
+        backtraceCommand[$backtraceNo]="${exception[$counter]}"
+        counter+=1
+        backtraceFile[$backtraceNo]="${exception[$counter]}"
+        counter+=1
+        
+        backtraceNo+=1
+    done
+    
+    local -i index=1
+    
+    while [[ $index -lt $backtraceNo ]]
+    do
+        # Log.Write "${traceFile[$index]}" \| "${traceLine[$index]}" \| "${traceCommand[($index - 1)]}" \| $(($index + 1))
+        Log.Write "$(Exception.FormatExceptionSegment "${backtraceFile[$index]}" "${backtraceLine[$index]}" "${backtraceCommand[($index - 1)]}" $(( $index + $backtraceLevel )) )"
+        index+=1
+    done
+}
+
+String.ReplaceSlashes() {
+    @var stringToMark
+    
+    # Workaround for a Bash bug that causes string replacement to fail when a \ is in the string
+    local slash="\\"
+    local slashReplacement='_%SLASH%_'
+    echo "${stringToMark/$slash$slash/$slashReplacement}"
+}
+
+String.BringBackSlashes() {
+    @var stringToMark
+    
+    # Workaround for a Bash bug that causes string replacement to fail when a \ is in the string
+    local slash="\\"
+    local slashReplacement='_%SLASH%_'
+    echo "${stringToMark/$slashReplacement/$slash}"
+}
+
+Exception.CanHighlight() {
+    @var errLine
+    @var stringToMark
+    
+    local stringToMarkWithoutSlash="$(String.ReplaceSlashes "$stringToMark")"
+    errLine="$(String.ReplaceSlashes "$errLine")"
+    
+    if [[ "$errLine" == *"$stringToMarkWithoutSlash"* ]]
+    then
+        return 0
+    else
+        return 1
+    fi
+}
+
+Exception.HighlightPart() {
+    @var errLine
+    @var stringToMark
+    
+    # Workaround for a Bash bug that causes string replacement to fail when a \ is in the string
+    local stringToMarkWithoutSlash="$(String.ReplaceSlashes "$stringToMark")"
+    errLine="$(String.ReplaceSlashes "$errLine")"
+    
+    local underlinedObject="$(Exception.GetUnderlinedPart "$stringToMark")"
+    local underlinedObjectInLine="${errLine/$stringToMarkWithoutSlash/$underlinedObject}"
+
+    # Bring back the slash:
+    underlinedObjectInLine="$(String.BringBackSlashes "$underlinedObjectInLine")"
+    
+    # Trimming:
+    underlinedObjectInLine="${underlinedObjectInLine#"${underlinedObjectInLine%%[![:space:]]*}"}" # "
+
+	echo "$underlinedObjectInLine"
+}
+
+Exception.GetUnderlinedPart() {
+    @var stringToMark
+    
+    echo "$(UI.Color.LightGreen)$(UI.Powerline.RefersTo) $(UI.Color.Magenta)$(UI.Color.Underline)$stringToMark$(UI.Color.White)$(UI.Color.NoUnderline)"
+}
+
+Exception.FormatExceptionSegment() {
     @var script
     @int lineNo
     @var stringToMark
     @int callPosition=1
 
     local errLine="$(sed "${lineNo}q;d" "$script")"
-
-    # Workaround for a Bash bug that causes string replacement to fail when a \ is in the string
-    local slash="\\"
-    local slashReplacement='_%SLASH%_'
-    local stringToMarkWithoutSlash="${stringToMark/$slash$slash/$slashReplacement}"
-    errLine="${errLine/$slash$slash/$slashReplacement}"
-
-    local underlinedObject="$(UI.Color.LightGreen)$(UI.Powerline.RefersTo) $(UI.Color.Magenta)$(UI.Color.Underline)$stringToMark$(UI.Color.White)$(UI.Color.NoUnderline)"
-    local underlinedObjectInLine="${errLine/$stringToMarkWithoutSlash/$underlinedObject}"
-
-    # Bring back the slash:
-    underlinedObjectInLine="${underlinedObjectInLine/$slashReplacement/$slash}"
-    underlinedObjectInLine="${underlinedObjectInLine#"${underlinedObjectInLine%%[![:space:]]*}"}" # trimming "
+    local originalErrLine="$errLine"
+    
+    local -i linesTried=0
+    
+    while [[ $linesTried -lt 5 && $lineNo -ge 0 ]] && ! Exception.CanHighlight "$errLine" "$stringToMark"
+    do
+        linesTried+=1
+        lineNo+=-1
+        errLine="$(sed "${lineNo}q;d" "$script")"
+    done
 
     # Cut out the path, leave the script name
     script="${script##*/}"
-
+    
     local prefix="   $(UI.Powerline.Branch)$(String.GetXSpaces $(($callPosition * 3 - 3)) || true) "
-    if [[ ! "$errLine" == *"$stringToMarkWithoutSlash"* ]]
+    
+    if [[ $linesTried -ge 5 ]]
     then
+        # PRINT THE ORGINAL OBJECT AND ORIGINAL LINE #
+        #local underlinedObject="$(Exception.HighlightPart "$errLine" "$stringToMark")"
+        local underlinedObject="$(Exception.GetUnderlinedPart "$stringToMark")"
         echo "${prefix}$(UI.Color.White)${underlinedObject}$(UI.Color.Default) [$(UI.Color.Blue)${script}:${lineNo}$(UI.Color.Default)]"
         prefix="$prefix$(UI.Powerline.Fail) "
+        errLine="$originalErrLine"
     fi
+    
+    local underlinedObjectInLine="$(Exception.HighlightPart "$errLine" "$stringToMark")"
+    
     echo "${prefix}$(UI.Color.White)${underlinedObjectInLine}$(UI.Color.Default) [$(UI.Color.Blue)${script}:${lineNo}$(UI.Color.Default)]"
 }
 
@@ -134,6 +253,31 @@ Exception.ContinueOrBreak()
     fi
 }
 
+Exception.DumpBacktrace()
+{
+    @int startFrom=1
+    Log.Write START FROM $startFrom vs $1
+    
+    # inspired by: http://stackoverflow.com/questions/64786/error-handling-in-bash
+    
+    local -i i=0
+
+    while caller $i > /dev/null
+    do
+        if (( $i + 1 >= $startFrom ))
+        then
+            #Log.Write CALLER: "'$(caller $i)'"
+            
+            local -a trace=( $(caller $i) )
+            
+            echo "${trace[0]}"
+            echo "${trace[1]}"
+            echo "${trace[@]:2}"
+        fi
+        i+=1
+    done
+}
+
 Exception.GetBacktrace()
 {
     @int startFrom
@@ -142,7 +286,6 @@ Exception.GetBacktrace()
     @reference filename
     @reference element
 
-    # inspired by: http://stackoverflow.com/questions/64786/error-handling-in-bash
 
     local -i i=0
